test-requirements.txt
nose
jsonschema
requests-mock
